load tests into one file and inherit from the imagenet task and the functioning tfds decoder for imagenet
