+++ training_dir/params.yaml
@@ -6,10 +6,10 @@ runtime:
   distribution_strategy: mirrored
   enable_xla: false
   gpu_thread_mode: null
-  loss_scale: dynamic
-  mixed_precision_dtype: float16
+  loss_scale: null
+  mixed_precision_dtype: float32
   num_cores_per_replica: 1
-  num_gpus: 2
+  num_gpus: 0
   num_packs: 1
   per_gpu_thread_count: 0
   run_eagerly: false
@@ -46,19 +46,19 @@ task:
     drop_remainder: true
     dtype: float16
     enable_tf_data_service: false
-    global_batch_size: 16
-    input_path: ''
+    global_batch_size: 128
+    input_path: imagenet-2012-tfrecord/train*
     is_training: true
     sharding: true
-    shuffle_buffer_size: 100
+    shuffle_buffer_size: 10000
     tf_data_service_address: null
     tf_data_service_job_name: null
     tfds_as_supervised: false
-    tfds_data_dir: ~/tensorflow_datasets/
-    tfds_download: true
-    tfds_name: imagenet2012
+    tfds_data_dir: ''
+    tfds_download: false
+    tfds_name: ''
     tfds_skip_decoding_feature: ''
-    tfds_split: train
+    tfds_split: ''
   validation_data:
     block_length: 1
     cache: false
@@ -67,19 +67,19 @@ task:
     drop_remainder: false
     dtype: float16
     enable_tf_data_service: false
-    global_batch_size: 16
-    input_path: ''
+    global_batch_size: 128
+    input_path: imagenet-2012-tfrecord/valid*
     is_training: true
     sharding: true
-    shuffle_buffer_size: 100
+    shuffle_buffer_size: 10000
     tf_data_service_address: null
     tf_data_service_job_name: null
     tfds_as_supervised: false
-    tfds_data_dir: ~/tensorflow_datasets/
-    tfds_download: true
-    tfds_name: imagenet2012
+    tfds_data_dir: ''
+    tfds_download: false
+    tfds_name: ''
     tfds_skip_decoding_feature: ''
-    tfds_split: validation
+    tfds_split: ''
 trainer:
   allow_tpu_summary: false
   best_checkpoint_eval_metric: ''
@@ -94,9 +94,9 @@ trainer:
     learning_rate:
       polynomial:
         cycle: false
-        decay_steps: 6392000
-        end_learning_rate: 1.25e-05
-        initial_learning_rate: 0.0125
+        decay_steps: 799000
+        end_learning_rate: 0.0001
+        initial_learning_rate: 0.1
         name: PolynomialDecay
         power: 4.0
       type: polynomial
@@ -113,12 +113,12 @@ trainer:
       linear:
         name: linear
         warmup_learning_rate: 0
-        warmup_steps: 8000
+        warmup_steps: 1000
       type: linear
   steps_per_loop: 10000
   summary_interval: 10000
-  train_steps: 6400000
+  train_steps: 800000
   train_tf_function: true
   train_tf_while_loop: true
   validation_interval: 10000
-  validation_steps: 3200
+  validation_steps: 400
